[feature] handle F with NaN
The current dff estimation crashes if there is a NaN in F. This can happen if one mask is defined outside of the part that suite2p can register(if one provide masks externally).
In this case the F is only NaN. Ensure that in this case dff is also NaN

If a line of f has some nan, but not all, the function will still crash.